feat: improved tools and decoupled them
