+++ backend/utils/gpt_wrapper.py
@@ -2,15 +2,12 @@ import json
 import os
 
 import dotenv
-from command_executor import CommandExecutor
+from gpt_tools import available_tools, tools
 from openai import OpenAI
-from supabase_wrapper import SupabaseWrapper
 
 dotenv.load_dotenv()
 
 client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
-supabase = SupabaseWrapper().client
-command_executor = CommandExecutor()
 
 
 system_prompt = """
@@ -31,6 +28,7 @@ You can decline to run a command if you think it's harmful.
 command_result_prompt = """
 You are a summarizer who summarizes the command stdout results.
 Say it if the command was successful or not.
+Print the stdout formatted as readable text.
 And what was the error if there was any.
 """
 
@@ -39,125 +37,34 @@ You are a summarizer who summarizes the command steps.
 Say how can the user do the steps from the command on the gui and on the command line.
 """
 
-
-tools = [
-    {
-        "type": "function",
-        "function": {
-            "name": "run_command",
-            "description": "Runs a command on the user's computer.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "command": {
-                        "type": "string",
-                        "description": "The command to run.",
-                    },
-                },
-            },
-            "required": ["command"],
-        },
-    },
-    {
-        "type": "function",
-        "function": {
-            "name": "run_workflow",
-            "description": "Runs multiple commands on the user's computer.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "commands": {
-                        "type": "array",
-                        "description": "The commands to run.",
-                        "items": {
-                            "type": "string",
-                        },
-                    },
-                },
-            },
-        },
-        "required": ["commands"],
-    },
-    {
-        "type": "function",
-        "function": {
-            "name": "get_workflows",
-            "description": "Gets a list of workflows.",
-            "parameters": {},
-        },
-    },
-    {
-        "type": "function",
-        "function": {
-            "name": "save_workflow",
-            "description": "Saves a workflow. The previously ran commands",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "title": {
-                        "type": "string",
-                        "description": "The title of the workflow.",
-                    },
-                    "description": {
-                        "type": "string",
-                        "description": "The description of the workflow.",
-                    },
-                    "definition": {
-                        "type": "string",
-                        "description": "The definition of the workflow. All the commands to run in a string.",
-                    },
-                },
-            },
-        },
-        "required": ["title", "description", "definition"],
-    },
-]
-
-
-def run_command(command):
-    try:
-        return command_executor.execute(command)
-    except Exception:
-        return "Failed to run command."
-
-
-def run_workflow(commands):
-    try:
-        return command_executor.execute_chain(commands)
-    except Exception:
-        return "Failed to run workflow."
-
-
-def get_workflows():
-    try:
-        return supabase.from_("Workflow").select("*").execute().data
-    except Exception:
-        return "Failed to get workflows."
-
-
-def save_workflow(title, description, definition):
-    try:
-        return (
-            supabase.from_("Workflow")
-            .insert(
-                {
-                    "title": title,
-                    "description": description,
-                    "definition": definition,
-                }
-            )
-            .execute()
-        )
-    except Exception:
-        return "Failed to save workflow."
-
-
-available_tools = {
-    "run_command": run_command,
-    "run_workflow": run_workflow,
-    "get_workflows": get_workflows,
-    "save_workflow": save_workflow,
-}
+db_summary_prompt = """
+Return the database results in a readable format.
+Expect a format like this:
+[
+  {
+    "id": 1,
+    "created_at": "2023-12-21T09:51:56.576832+00:00",
+    "title": "Test",
+    "description": "desc",
+    "definition": "ls -l",
+    "user_id": "bd7ab55b-590d-4b09-9a2a-5f682ae9acca"
+  },
+  {
+    "id": 2,
+    "created_at": "2023-12-21T10:01:47.671207+00:00",
+    "title": "Test2",
+    "description": "desc",
+    "definition": "echo hello",
+    "user_id": "bd7ab55b-590d-4b09-9a2a-5f682ae9acca"
+  },
+You should say the title and the definition of each workflow and the description if
+there is any even if they are in a list or dictionary.
+Don't say this "It returned a list of dictionaries with 'id', 'created_at', 'title', 'description', 'definition',
+and 'user_id' keys. Each dictionary corresponds to a record in the database."
+      }"
+Say the concrete values of the keys.
+Return a table with the columns 'id', 'title', 'description', and 'definition'.
+"""
 
 
 class GPTWrapper:
@@ -169,6 +76,7 @@ class GPTWrapper:
         self.system_prompt = system_prompt
         self.command_steps_prompt = command_steps_prompt
         self.command_result_prompt = command_result_prompt
+        self.db_summary_prompt = db_summary_prompt
         self.chat_history = []
         self.tools = tools
         self.append_system_prompt(system_prompt)
@@ -179,7 +87,6 @@ class GPTWrapper:
             model="gpt-3.5-turbo-1106", messages=self.chat_history, tools=self.tools, tool_choice="auto"
         )
         response = response.choices[0].message
-        # print(response)
         if response.tool_calls is not None:
             self.use_tools(response.tool_calls)
         else:
@@ -215,7 +122,6 @@ class GPTWrapper:
         return self.chat_history
 
     def use_tools(self, tool_calls):
-        # logger.info(tool_calls)
         for tool_call in tool_calls:
             function_name = tool_call.function.name
             function_to_call = available_tools[function_name]
@@ -227,8 +133,34 @@ class GPTWrapper:
                 "content": str(function_response),
             }
 
-            if (function_args is not None) and (function_args != {}):
-                self.append_assistant_message("Running command with parameters: " + str(function_args))
+            self.append_assistant_message(f"Running command {function_name} with parameters: {function_args}.")
+            self.append_assistant_message(f"Command result: {function_response}.")
+
+            if function_name == "get_workflows":
+                self.append_assistant_message("Getting workflows from database.")
+                summarize_function_calls = client.chat.completions.create(
+                    model="gpt-3.5-turbo-1106",
+                    messages=[
+                        {
+                            "role": "system",
+                            "content": self.db_summary_prompt,
+                        },
+                        tool_call_result,
+                    ],
+                )
+            if function_name == "get_workflow_from_db":
+                self.append_assistant_message(f"Getting workflow {function_args} from database.")
+                summarize_function_calls = client.chat.completions.create(
+                    model="gpt-3.5-turbo-1106",
+                    messages=[
+                        {
+                            "role": "system",
+                            "content": self.db_summary_prompt,
+                        },
+                        tool_call_result,
+                    ],
+                )
+            else:
                 summarize_function_calls = client.chat.completions.create(
                     model="gpt-3.5-turbo-1106",
                     messages=[
@@ -255,6 +187,6 @@ class GPTWrapper:
 
 
 if __name__ == "__main__":
-    gpt_wrapper = GPTWrapper()
-    gpt_wrapper.send_message("Hello, which workflows do i have?")
-    print(gpt_wrapper.chat_history)
+    gpt = GPTWrapper()
+    gpt.send_message("list my files on the desktop")
+    print(gpt.get_chat_history())
